requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
tensorflow==2.16.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
gunicorn==22.0.0
